+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Output quantification CSV as zip with timestamp notification
+- Update dependency [`uv`](https://github.com/camlab-bioml/rakaia/security/dependabot/236)
+- Increase default spatial radius size to 6
 
 ## [0.31.1] - 2026-05-20